hng_app/requirements.txt
Flask
requests
python-dotenv==0.19.2